fix: fix waitlist layout issue
