--- src/app/[locale]/waitlist/layout.tsx
@@ -1,12 +1,0 @@
-import { getWaitlistUser } from './helpers';
-import WaitlistPage from './page';
-
-export default async function WaitlistPageLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  const users = await getWaitlistUser();
-
-  return <WaitlistPage users={users} />;
-}
